requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
distlib==0.3.2
Django==3.2.6
djangorestframework==3.12.4
filelock==3.0.12
importlib-metadata==4.6.4
platformdirs==2.2.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
virtualenv==20.7.2
zipp==3.5.0
